% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{build.sdpd.model}
\alias{build.sdpd.model}
\title{Build an SDP-D Model Object}
\usage{
build.sdpd.model(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=build_sdpd_model]{build_sdpd_model()}}.}
}
\value{
See \code{\link[=build_sdpd_model]{build_sdpd_model()}}.
}
\description{
Deprecated wrapper for \code{\link[=build_sdpd_model]{build_sdpd_model()}}.
}
\seealso{
\code{\link[=build_sdpd_model]{build_sdpd_model()}}
}
