% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{build.spatial.matrix}
\alias{build.spatial.matrix}
\title{Build a Spatial Weight Matrix}
\usage{
build.spatial.matrix(ww.index, ww.values)
}
\arguments{
\item{ww.index}{Matrix. Spatial-neighbor index matrix.}

\item{ww.values}{Matrix. Spatial-weight value matrix.}
}
\value{
See \code{\link[=build_spatial_matrix]{build_spatial_matrix()}}.
}
\description{
Deprecated wrapper for \code{\link[=build_spatial_matrix]{build_spatial_matrix()}}.
}
\seealso{
\code{\link[=build_spatial_matrix]{build_spatial_matrix()}}
}
