% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{check.sdpd.model}
\alias{check.sdpd.model}
\title{Check an SDP-D Model}
\usage{
check.sdpd.model(
  res.fit = NULL,
  model = NULL,
  ww.index = NULL,
  ww.values = NULL
)
}
\arguments{
\item{res.fit}{Optional fitted SDP-D model object.}

\item{model}{Optional SDP-D model object.}

\item{ww.index}{Optional spatial-neighbor index matrix.}

\item{ww.values}{Optional spatial-weight value matrix.}
}
\value{
See \code{\link[=check_sdpd_model]{check_sdpd_model()}}.
}
\description{
Deprecated wrapper for \code{\link[=check_sdpd_model]{check_sdpd_model()}}.
}
\seealso{
\code{\link[=check_sdpd_model]{check_sdpd_model()}}
}
