% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fit.2nd.stage}
\alias{fit.2nd.stage}
\title{Fit the Second-Stage SDP-D Estimator}
\usage{
fit.2nd.stage(dseries, X.centr, W, model, coeff.hat, data = NULL)
}
\arguments{
\item{dseries}{Numeric matrix. Endogenous data series.}

\item{X.centr}{Numeric matrix or array of centered regressors.}

\item{W}{Numeric spatial weight matrix.}

\item{model}{SDP-D model object.}

\item{coeff.hat}{Numeric matrix of coefficient estimates.}

\item{data}{Optional data summary object.}
}
\value{
See \code{\link[=fit_second_stage]{fit_second_stage()}}.
}
\description{
Deprecated wrapper for \code{\link[=fit_second_stage]{fit_second_stage()}}.
}
\seealso{
\code{\link[=fit_second_stage]{fit_second_stage()}}
}
