% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fit.sdpd.assemble}
\alias{fit.sdpd.assemble}
\title{Assemble SDP-D Estimation Results}
\usage{
fit.sdpd.assemble(obj.stime)
}
\arguments{
\item{obj.stime}{List. Single spatial-time estimation object.}
}
\value{
See \code{\link[=fit_sdpd_assemble]{fit_sdpd_assemble()}}.
}
\description{
Deprecated wrapper for \code{\link[=fit_sdpd_assemble]{fit_sdpd_assemble()}}.
}
\seealso{
\code{\link[=fit_sdpd_assemble]{fit_sdpd_assemble()}}
}
