% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fit.sdpd.mean.equation.model}
\alias{fit.sdpd.mean.equation.model}
\title{Compute the SDP-D Mean-Equation Model}
\usage{
fit.sdpd.mean.equation.model(
  res,
  WW,
  time.weights = NULL,
  index.weights = NULL
)
}
\arguments{
\item{res}{List. SDP-D estimation result object.}

\item{WW}{Numeric matrix. Spatial weight matrix.}

\item{time.weights}{Optional factor or numeric vector of time weights.}

\item{index.weights}{Optional numeric vector of spatial-unit weights.}
}
\value{
See \code{\link[=fit_sdpd_mean_equation_model]{fit_sdpd_mean_equation_model()}}.
}
\description{
Deprecated wrapper for \code{\link[=fit_sdpd_mean_equation_model]{fit_sdpd_mean_equation_model()}}.
}
\seealso{
\code{\link[=fit_sdpd_mean_equation_model]{fit_sdpd_mean_equation_model()}}
}
