% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fun.JBtest}
\alias{fun.JBtest}
\title{Compute a Jarque-Bera Normality-Test P-Value}
\usage{
fun.JBtest(x, ...)
}
\arguments{
\item{x}{Numeric vector. Input values to test for normality.}

\item{...}{Additional arguments passed to \code{\link[=fun_jb_test]{fun_jb_test()}}.}
}
\value{
See \code{\link[=fun_jb_test]{fun_jb_test()}}.
}
\description{
Deprecated wrapper for \code{\link[=fun_jb_test]{fun_jb_test()}}.
}
\seealso{
\code{\link[=fun_jb_test]{fun_jb_test()}}
}
