% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{fun.LBtest}
\alias{fun.LBtest}
\title{Compute a Ljung-Box Test P-Value}
\usage{
fun.LBtest(x, ...)
}
\arguments{
\item{x}{Numeric vector. Input values to test for serial autocorrelation.}

\item{...}{Additional arguments passed to \code{\link[=fun_lb_test]{fun_lb_test()}}.}
}
\value{
See \code{\link[=fun_lb_test]{fun_lb_test()}}.
}
\description{
Deprecated wrapper for \code{\link[=fun_lb_test]{fun_lb_test()}}.
}
\seealso{
\code{\link[=fun_lb_test]{fun_lb_test()}}
}
