% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{generate.sdpd.model}
\alias{generate.sdpd.model}
\title{Generate an SDP-D Model}
\usage{
generate.sdpd.model(pp, model, ww.index, ww.values, px.neighbors, sim.options)
}
\arguments{
\item{pp}{Integer. Number of spatial units.}

\item{model}{SDP-D model object.}

\item{ww.index}{Optional spatial-neighbor index matrix.}

\item{ww.values}{Optional spatial-weight value matrix.}

\item{px.neighbors}{Optional proximity-neighbor structure.}

\item{sim.options}{List. Simulation options.}
}
\value{
See \code{\link[=generate_sdpd_model]{generate_sdpd_model()}}.
}
\description{
Deprecated wrapper for \code{\link[=generate_sdpd_model]{generate_sdpd_model()}}.
}
\seealso{
\code{\link[=generate_sdpd_model]{generate_sdpd_model()}}
}
