% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{HSDPD-package}
\alias{HSDPD}
\alias{HSDPD-package}
\title{HSDPD: Hierarchical Spatial Dynamic Panel Data Models}
\description{
Tools for building, simulating, estimating, testing, and plotting
hierarchical spatial dynamic panel data models.
}
\author{
\strong{Maintainer}: Michele La Rocca \email{larocca@unisa.it}

Authors:
\itemize{
  \item Michele La Rocca \email{larocca@unisa.it}
  \item Maria Lucia Parrella \email{mparrel@unisa.it}
}

Other contributors:
\itemize{
  \item Francesco Giordano \email{giordano@unisa.it} [contributor]
  \item Cira Perna \email{perna@unisa.it} [contributor]
}

}
\keyword{internal}
