% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{parallelize.sdpd.data}
\alias{parallelize.sdpd.data}
\title{Parallelize SDP-D Series Data by Group}
\usage{
parallelize.sdpd.data(obj.series, model)
}
\arguments{
\item{obj.series}{SDP-D series object.}

\item{model}{SDP-D model object.}
}
\value{
See \code{\link[=parallelize_sdpd_data]{parallelize_sdpd_data()}}.
}
\description{
Deprecated wrapper for \code{\link[=parallelize_sdpd_data]{parallelize_sdpd_data()}}.
}
\seealso{
\code{\link[=parallelize_sdpd_data]{parallelize_sdpd_data()}}
}
