% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot.sdpd.boot.series}
\alias{plot.sdpd.boot.series}
\title{Plot a Bootstrap SDP-D Series}
\usage{
\method{plot}{sdpd.boot.series}(bb, pp, nn, eigen, yy.star1, opts.boot, res.fit, group.index)
}
\arguments{
\item{bb}{Integer. Current bootstrap iteration.}

\item{pp}{Integer. Number of spatial units.}

\item{nn}{Integer. Number of time observations.}

\item{eigen}{Numeric vector. Eigenvalue diagnostics.}

\item{yy.star1}{Numeric matrix. Bootstrapped time series.}

\item{opts.boot}{List. Bootstrap options.}

\item{res.fit}{List. Fitted SDP-D model result.}

\item{group.index}{Vector. Group index.}
}
\value{
Invisibly returns the output file path of the generated JPEG plot.
}
\description{
Deprecated wrapper for \code{\link[=plot_sdpd_boot_series]{plot_sdpd_boot_series()}}.
}
\seealso{
\code{\link[=plot_sdpd_boot_series]{plot_sdpd_boot_series()}}
}
