% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot.sdpd.estimates}
\alias{plot.sdpd.estimates}
\title{Plot Spatial SDP-D Estimates}
\usage{
\method{plot}{sdpd.estimates}(
  obj.results,
  item = "coeff.hat",
  main = NULL,
  sub = NULL,
  limits = NULL,
  mid = 0,
  size.point = 1
)
}
\arguments{
\item{obj.results}{List or data frame. SDP-D estimation results.}

\item{item}{Character scalar. Name of the result component to plot.}

\item{main}{Optional character scalar. Text prepended to each plot title.}

\item{sub}{Optional character scalar. Text appended to each plot title.}

\item{limits}{Optional numeric vector of length 2.}

\item{mid}{Numeric scalar or vector. Midpoint value for the diverging color
scale.}

\item{size.point}{Numeric scalar. Point size used in the plot.}
}
\value{
See \code{\link[=plot_sdpd_estimates]{plot_sdpd_estimates()}}.
}
\description{
Deprecated wrapper for \code{\link[=plot_sdpd_estimates]{plot_sdpd_estimates()}}.
}
\seealso{
\code{\link[=plot_sdpd_estimates]{plot_sdpd_estimates()}}
}
