% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot.sdpd.model}
\alias{plot.sdpd.model}
\title{Plot an SDP-D Model Fit}
\usage{
\method{plot}{sdpd.model}(
  res.fit,
  n.units = "all",
  n.vars = "all",
  which = c(1, 2)[1],
  t.axis = list(t.labels = NULL, t.points = NULL),
  xlimit = NULL,
  ylimit = NULL,
  max.col = 5,
  col.punti = 1
)
}
\arguments{
\item{res.fit}{Fitted SDP-D model result object.}

\item{n.units}{Integer or \code{"all"}. Number of spatial units to plot.}

\item{n.vars}{Integer or \code{"all"}. Number of variables to plot.}

\item{which}{Integer scalar. Plot type selector.}

\item{t.axis}{List. Time-axis options with elements \code{t.labels} and
\code{t.points}.}

\item{xlimit}{Optional numeric vector of length 2. X-axis limits.}

\item{ylimit}{Optional numeric vector of length 2. Y-axis limits.}

\item{max.col}{Integer. Maximum number of plot columns.}

\item{col.punti}{Color specification for plotted points or lines.}
}
\value{
See \code{\link[=plot_sdpd_model]{plot_sdpd_model()}}.
}
\description{
Deprecated wrapper for \code{\link[=plot_sdpd_model]{plot_sdpd_model()}}.
}
\seealso{
\code{\link[=plot_sdpd_model]{plot_sdpd_model()}}
}
