% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{read.data.from.raster}
\alias{read.data.from.raster}
\title{Read SDP-D Data from Raster Objects}
\usage{
read.data.from.raster(
  px = NULL,
  latit = NULL,
  longit = NULL,
  rry,
  rrXX = NULL,
  rrgroups = NULL,
  label_groups = NULL,
  model,
  vec.options,
  type.w = "distance"
)
}
\arguments{
\item{px}{Optional vector. Raster cell identifiers to include.}

\item{latit}{Optional latitude vector.}

\item{longit}{Optional longitude vector.}

\item{rry}{\code{SpatRaster}. Endogenous raster time series.}

\item{rrXX}{Optional \code{SpatRaster} or list of \code{SpatRaster} objects.}

\item{rrgroups}{Optional group raster.}

\item{label_groups}{Optional named vector of group labels.}

\item{model}{SDP-D model object.}

\item{vec.options}{List. Vectorization options.}

\item{type.w}{Character scalar. Spatial-weight type.}
}
\value{
See \code{\link[=read_data_from_raster]{read_data_from_raster()}}.
}
\description{
Deprecated wrapper for \code{\link[=read_data_from_raster]{read_data_from_raster()}}.
}
\seealso{
\code{\link[=read_data_from_raster]{read_data_from_raster()}}
}
