% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{test.sdpd.model}
\alias{test.sdpd.model}
\title{Test an SDP-D Model by Residual Bootstrap}
\usage{
test.sdpd.model(
  res.fit,
  px,
  model = NULL,
  n.boot = 399,
  H0 = c("zero", "constant", "grouped", "nospatial", "noautoregressive", "noX",
    "constrained")[1],
  coeff.H0 = 0,
  group.index = rep(1, dim(res.fit$fitted)[1]),
  opts.boot = list(markovian = FALSE, resid = c("fitted", "normal")[1], sigma.resid = 1,
    boot.plot = FALSE, cartel = "", ylimiti = NULL, label.index = NULL)
)
}
\arguments{
\item{res.fit}{Fitted SDP-D model result object.}

\item{px}{Vector. Spatial-unit identifiers.}

\item{model}{Optional SDP-D model object.}

\item{n.boot}{Integer. Number of bootstrap replications.}

\item{H0}{Character or integer. Null hypothesis to test.}

\item{coeff.H0}{Numeric scalar or matrix. Coefficient value under the null.}

\item{group.index}{Vector. Group index.}

\item{opts.boot}{List. Bootstrap options.}
}
\value{
See \code{\link[=test_sdpd_model]{test_sdpd_model()}}.
}
\description{
Deprecated wrapper for \code{\link[=test_sdpd_model]{test_sdpd_model()}}.
}
\seealso{
\code{\link[=test_sdpd_model]{test_sdpd_model()}}
}
